REFACTOR: Changed player renderable into an entity
