+++ src/index.ts
@@ -1,7 +1,7 @@
 import { audio, loader, state, device, video, plugin, pool, input } from "melonjs";
 import TitleScreen from "./scripts/stage/title.js";
 import PlayScreen from "./scripts/stage/play.js";
-import PlayerEntity from "./scripts/renderables/player.js";
+import PlayerEntity from "./scripts/entities/player.js";
 import DataManifest from "./manifest.js";
 import "./index.scss"
