+++ src/scripts/entities/player.ts
@@ -1,0 +1,125 @@
+import * as me from "melonjs";
+import ResponseObject from "melonjs/dist/types/physics/response";
+
+class PlayerEntity extends me.Entity {
+
+    public static PLAYER_HEIGHT = 64;
+    public static PLAYER_WIDTH = 48;
+    public static JUMP_VELOCITY = -20;
+    public static DUCK_HEIGHT = 32;
+
+    private gravity: number = 0.9;
+    private maxFallSpeed: number = 20;
+
+    private isJumping: boolean = false;
+    private isDucking: boolean = false;
+    private isGrounded: boolean = false;
+
+    private groundY: number = 0;
+
+    constructor(x: number, groundY: number) {
+        super(x / 2, groundY / 2, { width: PlayerEntity.PLAYER_WIDTH, height: PlayerEntity.PLAYER_HEIGHT, name: "player" });
+        this.groundY = groundY / 2;
+
+        this.body.setMaxVelocity(0, this.maxFallSpeed);
+        this.body.setFriction(0, 0);
+
+        this.body.setCollisionMask(me.collision.types.ALL_OBJECT);
+        this.body.collisionType = me.collision.types.PLAYER_OBJECT;
+
+        this.body.addShape(new me.Rect(0, 0, this.width, this.height));
+        this.body.ignoreGravity = true
+        this.alwaysUpdate = true;
+
+        this.anchorPoint.set(0.5, 0);
+    }
+
+    private jump() {
+        if (this.body.vel.y === 0 && !this.isDucking) {
+            this.isJumping = true;
+            this.body.vel.y = PlayerEntity.JUMP_VELOCITY;
+        }
+    }
+
+    private duckStart() {
+        this.isDucking = true;
+
+        this.height = PlayerEntity.DUCK_HEIGHT;
+        this.body.removeShapeAt(0);
+        const shape = new me.Rect(0, 0, this.width, this.height);
+        this.body.addShape(shape);
+        this.pos.y! += PlayerEntity.PLAYER_HEIGHT - PlayerEntity.DUCK_HEIGHT;
+    }
+
+    private duckEnd() {
+        this.isDucking = false;
+
+        this.height = PlayerEntity.PLAYER_HEIGHT;
+        this.body.removeShapeAt(0);
+        const shape = new me.Rect(0, 0, this.width, this.height);
+        this.body.addShape(shape);
+        this.pos.y! += PlayerEntity.PLAYER_HEIGHT + PlayerEntity.DUCK_HEIGHT;
+    }
+
+    public update(dt: number): boolean {
+        if (me.input.isKeyPressed("jump")) {
+            this.jump();
+        }
+        if (me.input.isKeyPressed("duck")) {
+            if (!this.isJumping && !this.isDucking) {
+                this.duckStart();
+            }
+        } else if (this.isDucking) {
+            this.duckEnd();
+        }
+
+        if (this.body.vel.y !== 0) {
+            this.body.vel.y! += this.gravity;
+        }
+
+        if (this.pos.y! > this.groundY - this.height / 2) {
+            this.isJumping = false;
+            this.pos.y = this.groundY - this.height / 2;
+            this.body.vel.y = 0;
+        } else if (this.pos.y! < 1) {
+            this.body.vel.y = 1;
+        }
+
+        return true;
+    }
+
+    // onCollision(response: ResponseObject): boolean {
+    //     if (response.b.body.collisionType === me.collision.types.ENEMY_OBJECT) {
+    //         return false;
+    //     }
+
+    //     return false;
+    // }
+
+    public draw(renderer: me.Renderer) {
+
+        let color = new me.Color(0, 255, 0);
+        if (this.isDucking) {
+            color.setColor(0, 0, 255);
+        } else if (this.isJumping) {
+            color.setColor(255, 0, 0);
+        }
+
+        renderer.currentColor = color;
+
+        const rect: me.Rect = new me.Rect(
+            this.pos.x ? this.pos.x : 0,
+            this.pos.y ? this.pos.y : 0,
+            this.width,
+            this.height
+        );
+
+        renderer.stroke(rect, true);
+    }
+
+    public destroy(...args: any[]): void {
+        super.destroy(...args);
+    }
+};
+
+export default PlayerEntity;
